update user with created Round
