+++ server/models/User.js
@@ -66,8 +66,6 @@ userSchema.methods.isCorrectPassword = async function(password) {
   userSchema.virtual('friendCount').get(function() {
     return this.friends.length;
   });
-  
-
 
 const User = model('User', userSchema);
