+++ server/seeders/seeds.js
@@ -1,7 +1,7 @@
 const db = require('../config/connection');
 const { User, Course, Round } = require('../models');
+const userData = require('./user-seeds');
 const courseData = require('./course-seeds');
-const { faker } = require('@faker-js/faker');
 
 db.once('open', async() => {
     await User.deleteMany({});
@@ -9,13 +9,6 @@ db.once('open', async() => {
     await Round.deleteMany({});
 
     // create users
-    const userData = [];
-    for (let i = 0; i < 10; i += 1) {
-        const username = faker.internet.userName();
-        const email = faker.internet.email(username);
-        const password = faker.internet.password();
-        userData.push({ username, email, password });
-    }
     await User.collection.insertMany(userData);
 
     // create courses
@@ -29,7 +22,6 @@ db.once('open', async() => {
     for (let i = 0; i < 10; i += 1) {
         const randomUserIndex = Math.floor(Math.random() * i) + 1;
         const username = generatedUsers[randomUserIndex];
-
         const randomCourseIndex = Math.floor(Math.random() * 4) + 1;
         const courseName = generatedCourses[randomCourseIndex];
 
@@ -39,12 +31,18 @@ db.once('open', async() => {
             const stroke = Math.floor(Math.random() * 7) + 1;
             scores.push({ holeNumber, stroke });
         }
-        roundData.push({username, courseName, scores: scores});
+        const createdRound = await Round.create({username, courseName, scores: scores});
+        
+        // add round data to user
+        await User.updateOne(
+            { username: username },
+            { $push: { rounds: createdRound._id }}
+        );
+        roundData.push(createdRound);
     }
-    await Round.collection.insertMany(roundData);
 
-    // add friends to users
     // add courses to users
+    // add friends to users
 
     console.log('Database has been seeded.');
     process.exit(0);
